+++ CHANGELOG.md
@@ -12,3 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Improved test coverage
 * Configuration via environment variables 
 
+## [0.2.0] - 04-12-2020
+### Added
+* Automatically determine package name
+
+### [0.3.0] - 17-12-2020
+### Added
+* Enable / disable public S3 ACL via command line
+